fix error in slice_contours

+++ Code/Applications/slice_contours.cxx
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]) {
 		("yellow,y", bpo::value<std::vector<std::string>	>(&ysurfs)->multitoken(), "surfaces to display in yellow color")
 		("green,g", bpo::value<std::vector<std::string>	>(&gsurfs)->multitoken(), "surfaces to display in green color")
 		("blue,b", bpo::value<std::vector<std::string>	>(&bsurfs)->multitoken(), "surfaces to display in blue color")
-		("num-slices,n", bpo::value < int >(&nimages)->default_value(14), "number of slices")
+		("num-slices,n", bpo::value < size_t >(&nimages)->default_value(14), "number of slices")
 		("slices,s", bpo::value < std::vector<int> >(&sl_vect)->multitoken(), "slice number")
 		("axis,a", bpo::value < std::vector<std::string> >(&axisnames), "axes to be extracted")
 		("all-axis,A", bpo::bool_switch(), "export all axes");
